refactor(server/parsers): Delete reddit, split 4chan into multiple files
